Move all the GUI objects into the gui module
