+++ gpgsync/gui/gpgsync.py
