+++ gpgsync/gui/settings_dialog.py
