+++ gpgsync/gui/systray.py
